`make html` for devs, `make versions` to deploy #22
We want `make html` to work fine for devs who are editing the docs and
want to preview changes. Devs or doc writers or whatever.

When we build the docs to deploy to production, we want all the versions
of the docs to be built. These are only built from tags or the commits
in branches. That's how the sphinx-multiversion extension works. To call
this purpose `make versions` has been added.